Fix remaining test file issues.
The remaining failures will be exempted.
