+++ app/src/sharedTest/java/org/oppia/android/app/help/faq/FAQListActivityTest.kt
@@ -1,21 +1,13 @@
-package org.oppia.android.app.testing.administratorcontrols
+package org.oppia.android.app.help.faq
 
 import android.app.Application
 import android.content.Context
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.RecyclerView
-import androidx.test.core.app.ActivityScenario.launch
+import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.contrib.RecyclerViewActions.scrollToPosition
-import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import dagger.Component
-import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -23,8 +15,6 @@ import org.junit.runner.RunWith
 import org.oppia.android.app.activity.ActivityComponent
 import org.oppia.android.app.activity.ActivityComponentFactory
 import org.oppia.android.app.activity.route.ActivityRouterModule
-import org.oppia.android.app.administratorcontrols.AdministratorControlsActivity
-import org.oppia.android.app.administratorcontrols.appversion.AppVersionFragment
 import org.oppia.android.app.application.ApplicationComponent
 import org.oppia.android.app.application.ApplicationInjector
 import org.oppia.android.app.application.ApplicationInjectorProvider
@@ -33,9 +23,8 @@ import org.oppia.android.app.application.ApplicationStartupListenerModule
 import org.oppia.android.app.application.testing.TestingBuildFlavorModule
 import org.oppia.android.app.devoptions.DeveloperOptionsModule
 import org.oppia.android.app.devoptions.DeveloperOptionsStarterModule
-import org.oppia.android.app.model.ProfileId
+import org.oppia.android.app.model.ScreenName
 import org.oppia.android.app.player.state.itemviewmodel.SplitScreenInteractionModule
-import org.oppia.android.app.settings.profile.ProfileListFragment
 import org.oppia.android.app.shim.ViewBindingShimModule
 import org.oppia.android.app.test.R
 import org.oppia.android.app.translation.testing.ActivityRecreatorTestModule
@@ -71,6 +60,7 @@ import org.oppia.android.domain.platformparameter.PlatformParameterModule
 import org.oppia.android.domain.platformparameter.PlatformParameterSingletonModule
 import org.oppia.android.domain.question.QuestionModule
 import org.oppia.android.domain.workmanager.WorkManagerConfigurationModule
+import org.oppia.android.testing.OppiaTestRule
 import org.oppia.android.testing.TestLogReportingModule
 import org.oppia.android.testing.firebase.TestAuthenticationModule
 import org.oppia.android.testing.junit.InitializeDefaultLocaleRule
@@ -83,6 +73,7 @@ import org.oppia.android.util.caching.AssetModule
 import org.oppia.android.util.caching.testing.CachingTestModule
 import org.oppia.android.util.gcsresource.GcsResourceModule
 import org.oppia.android.util.locale.LocaleProdModule
+import org.oppia.android.util.logging.CurrentAppScreenNameIntentDecorator.extractCurrentAppScreenName
 import org.oppia.android.util.logging.LoggerModule
 import org.oppia.android.util.logging.SyncStatusModule
 import org.oppia.android.util.logging.firebase.FirebaseLogUploaderModule
@@ -92,99 +83,66 @@ import org.oppia.android.util.parser.html.HtmlParserEntityTypeModule
 import org.oppia.android.util.parser.image.GlideImageLoaderModule
 import org.oppia.android.util.parser.image.ImageParsingModule
 import org.robolectric.annotation.Config
+import org.robolectric.annotation.LooperMode
 import javax.inject.Inject
 import javax.inject.Singleton
 
+/** Tests for [FAQListActivity]. */
 @RunWith(AndroidJUnit4::class)
-@Config(
-  application = AdministratorControlsFragmentTest.TestApplication::class,
-  qualifiers = "port-xxhdpi"
-)
-class AdministratorControlsFragmentTest {
-  @get:Rule
-  val initializeDefaultLocaleRule = InitializeDefaultLocaleRule()
+@LooperMode(LooperMode.Mode.PAUSED)
+@Config(application = FAQListActivityTest.TestApplication::class, qualifiers = "port-xxhdpi")
+class FAQListActivityTest {
+  @get:Rule val initializeDefaultLocaleRule = InitializeDefaultLocaleRule()
+  @get:Rule val oppiaTestRule = OppiaTestRule()
 
-  @Inject
-  lateinit var context: Context
-
-  @Inject
-  lateinit var testCoroutineDispatchers: TestCoroutineDispatchers
+  @Inject lateinit var context: Context
+  @Inject lateinit var testCoroutineDispatchers: TestCoroutineDispatchers
 
   @Before
   fun setUp() {
-    Intents.init()
     setUpTestApplicationComponent()
-    testCoroutineDispatchers.registerIdlingResource()
   }
 
-  @After
-  fun tearDown() {
-    testCoroutineDispatchers.unregisterIdlingResource()
-    Intents.release()
+  @Test
+  fun testActivity_createIntent_verifyScreenNameInIntent() {
+    val intent = FAQListActivity.createFAQListActivityIntent(context)
+
+    val screenName = intent.extractCurrentAppScreenName()
+    assertThat(screenName).isEqualTo(ScreenName.FAQ_LIST_ACTIVITY)
+  }
+
+  @Test
+  fun testFaqListActivity_hasCorrectActivityLabel() {
+    runWithLaunchedActivity {
+      onActivity { activity ->
+        val title = activity.title
+
+        // Verify that the activity label is correct as a proxy to verify TalkBack will announce the
+        // correct string when it's read out.
+        assertThat(title).isEqualTo(context.getString(R.string.faq_activity_title))
+      }
+    }
   }
 
   private fun setUpTestApplicationComponent() {
     ApplicationProvider.getApplicationContext<TestApplication>().inject(this)
   }
 
-  @Test
-  @Config(qualifiers = "sw600dp")
-  fun testAdministratorControlsFragment_clickEditProfile_checkLoadingTheCorrectFragment() {
-    launch<AdministratorControlsActivity>(
-      createAdministratorControlsActivityIntent(
-        0
-      )
-    ).use {
+  private fun runWithLaunchedActivity(testBlock: ActivityScenario<FAQListActivity>.() -> Unit) {
+    val intent = FAQListActivity.createFAQListActivityIntent(context)
+    ActivityScenario.launch<FAQListActivity>(intent).use { scenario ->
       testCoroutineDispatchers.runCurrent()
-      onView(withId(R.id.edit_profiles_text_view))
-        .perform(click())
-      it.onActivity { activity ->
-        val fragment =
-          activity.supportFragmentManager
-            .findFragmentById(R.id.administrator_controls_fragment_multipane_placeholder)
-        assertThat(fragment is ProfileListFragment).isTrue()
-      }
+      scenario.testBlock()
     }
-  }
-
-  @Test
-  @Config(qualifiers = "sw600dp")
-  fun testAdministratorControlsFragment_clickAppVersion_checkLoadingTheCorrectFragment() {
-    launch<AdministratorControlsActivity>(
-      createAdministratorControlsActivityIntent(
-        0
-      )
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(withId(R.id.administrator_controls_list)).perform(
-        scrollToPosition<RecyclerView.ViewHolder>(
-          3
-        )
-      )
-      onView(withId(R.id.app_version_text_view)).perform(click())
-      it.onActivity { activity ->
-        val fragment =
-          activity.supportFragmentManager
-            .findFragmentById(R.id.administrator_controls_fragment_multipane_placeholder)
-        assertThat(fragment is AppVersionFragment).isTrue()
-      }
-    }
-  }
-
-  private fun createAdministratorControlsActivityIntent(internalProifleId: Int): Intent {
-    val profileId = ProfileId.newBuilder().setInternalId(internalProifleId).build()
-    return AdministratorControlsActivity.createAdministratorControlsActivityIntent(
-      context,
-      profileId
-    )
   }
 
   // TODO(#59): Figure out a way to reuse modules instead of needing to re-declare them.
   @Singleton
   @Component(
     modules = [
-      TestDispatcherModule::class, ApplicationModule::class, RobolectricModule::class,
+      RobolectricModule::class,
       PlatformParameterModule::class, PlatformParameterSingletonModule::class,
+      TestDispatcherModule::class, ApplicationModule::class,
       LoggerModule::class, ContinueModule::class, FractionInputModule::class,
       ItemSelectionInputModule::class, MultipleChoiceInputModule::class,
       NumberWithUnitsRuleModule::class, NumericInputRuleModule::class, TextInputRuleModule::class,
@@ -217,18 +175,18 @@ class AdministratorControlsFragmentTest {
       override fun build(): TestApplicationComponent
     }
 
-    fun inject(testAdministratorControlsFragmentTest: AdministratorControlsFragmentTest)
+    fun inject(faqListActivityTest: FAQListActivityTest)
   }
 
   class TestApplication : Application(), ActivityComponentFactory, ApplicationInjectorProvider {
     private val component: TestApplicationComponent by lazy {
-      DaggerAdministratorControlsFragmentTest_TestApplicationComponent.builder()
+      DaggerFAQListActivityTest_TestApplicationComponent.builder()
         .setApplication(this)
         .build() as TestApplicationComponent
     }
 
-    fun inject(testAdministratorControlsFragmentTest: AdministratorControlsFragmentTest) {
-      component.inject(testAdministratorControlsFragmentTest)
+    fun inject(faqListActivityTest: FAQListActivityTest) {
+      component.inject(faqListActivityTest)
     }
 
     override fun createActivityComponent(activity: AppCompatActivity): ActivityComponent {
