+++ domain/src/test/java/org/oppia/android/domain/classify/rules/dragAndDropSortInput/DragDropSortInputHasElementXBeforeElementYClassifierProviderTest.kt
@@ -20,29 +20,28 @@ import org.robolectric.annotation.LooperMode
 import javax.inject.Inject
 import javax.inject.Singleton
 
-/** Tests for [DragDropSortInputHasElementXAtPositionYClassifierProvider]. */
+/** Tests for [DragDropSortInputHasElementXBeforeElementYClassifierProvider]. */
 @Suppress("PrivatePropertyName") // Truly immutable constants can be named in CONSTANT_CASE.
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
 @Config(manifest = Config.NONE)
-class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
+class DragDropSortInputHasElementXBeforeElementYClassifierProviderTest {
 
-  private val NON_NEGATIVE_VALUE_0 = createNonNegativeInt(value = 1)
-  private val NON_NEGATIVE_VALUE_1 = createNonNegativeInt(value = 2)
-  private val VALID_CONTENT_ID_2 = createTranslatableHtmlContentId(contentId = "valid_content_id_2")
+  private val VALID_CONTENT_ID_1 = createTranslatableHtmlContentId(contentId = "content_id_1")
+  private val VALID_CONTENT_ID_2 = createTranslatableHtmlContentId(contentId = "content_id_2")
   private val INVALID_CONTENT_ID = createTranslatableHtmlContentId(contentId = "invalid_content_id")
+  private val NON_NEGATIVE_VALUE_1 = createNonNegativeInt(value = 1)
   private val LIST_OF_SETS_OF_CONTENT_IDS =
     createListOfSetsOfTranslatableHtmlContentIds(
-      listOf("other_id_1", "other_id_2"),
-      listOf("valid_content_id_1", "valid_content_id_2", "valid_content_id_3")
+      listOf("content_id_1"), listOf("content_id_2"), listOf("content_id_3")
     )
 
   @Inject
-  internal lateinit var dragDropSortInputHasElementXAtPositionYClassifierProvider:
-    DragDropSortInputHasElementXAtPositionYClassifierProvider
+  internal lateinit var dragDropSortInputHasElementXBeforeElementYClassifierProvider:
+    DragDropSortInputHasElementXBeforeElementYClassifierProvider
 
-  private val hasElementXAtPositionYRuleClassifier: RuleClassifier by lazy {
-    dragDropSortInputHasElementXAtPositionYClassifierProvider.createRuleClassifier()
+  private val hasElementXBeforeElementYRuleClassifier: RuleClassifier by lazy {
+    dragDropSortInputHasElementXBeforeElementYClassifierProvider.createRuleClassifier()
   }
 
   @Before
@@ -51,12 +50,14 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
   }
 
   @Test
-  fun testAnswer_nonNegativeInput_testString_bothInputsWithIncorrectTypes_throwsException() {
-    // Reverse the x and y parameters to ensure both have the incorrect type.
-    val inputs = mapOf("x" to NON_NEGATIVE_VALUE_1, "y" to VALID_CONTENT_ID_2)
+  fun testAnswer_nonNegativeInput_bothInputsWithIncorrectTypes_throwsException() {
+    val inputs = mapOf(
+      "x" to NON_NEGATIVE_VALUE_1,
+      "y" to NON_NEGATIVE_VALUE_1
+    )
 
     val exception = assertThrows<IllegalStateException>() {
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -70,13 +71,10 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
 
   @Test
   fun testAnswer_nonNegativeInput_testString_xInputWithIncorrectType_throwsException() {
-    val inputs = mapOf(
-      "x" to NON_NEGATIVE_VALUE_1,
-      "y" to NON_NEGATIVE_VALUE_1
-    )
+    val inputs = mapOf("x" to NON_NEGATIVE_VALUE_1, "y" to VALID_CONTENT_ID_2)
 
     val exception = assertThrows<IllegalStateException>() {
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -90,10 +88,10 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
 
   @Test
   fun testAnswer_nonNegativeInput_testString_yInputWithIncorrectType_throwsException() {
-    val inputs = mapOf("x" to VALID_CONTENT_ID_2, "y" to VALID_CONTENT_ID_2)
+    val inputs = mapOf("x" to VALID_CONTENT_ID_2, "y" to NON_NEGATIVE_VALUE_1)
 
     val exception = assertThrows<IllegalStateException>() {
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -102,7 +100,7 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
 
     assertThat(exception)
       .hasMessageThat()
-      .contains("Expected input value to be of type NON_NEGATIVE_INT")
+      .contains("Expected input value to be of type TRANSLATABLE_HTML_CONTENT_ID")
   }
 
   @Test
@@ -110,7 +108,7 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
     val inputs = mapOf("y" to VALID_CONTENT_ID_2)
 
     val exception = assertThrows<IllegalStateException>() {
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -127,7 +125,7 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
     val inputs = mapOf("x" to VALID_CONTENT_ID_2)
 
     val exception = assertThrows<IllegalStateException>() {
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -144,7 +142,7 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
     val inputs = mapOf("z" to VALID_CONTENT_ID_2)
 
     val exception = assertThrows<IllegalStateException>() {
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -157,11 +155,11 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
   }
 
   @Test
-  fun testAnswer_elementXWithPositionY_bothValueDoNotMatch() {
-    val inputs = mapOf("y" to NON_NEGATIVE_VALUE_0, "x" to INVALID_CONTENT_ID)
+  fun testAnswer_elementXAfterElementY_orderedIncorrectly() {
+    val inputs = mapOf("x" to VALID_CONTENT_ID_2, "y" to VALID_CONTENT_ID_1)
 
     val matches =
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -171,11 +169,11 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
   }
 
   @Test
-  fun testAnswer_elementXWithPositionY_xValueDoesNotMatch() {
-    val inputs = mapOf("y" to NON_NEGATIVE_VALUE_1, "x" to INVALID_CONTENT_ID)
+  fun testAnswer_elementX_invalidElementY_orderedIncorrectly() {
+    val inputs = mapOf("y" to INVALID_CONTENT_ID, "x" to VALID_CONTENT_ID_2)
 
     val matches =
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -185,25 +183,11 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
   }
 
   @Test
-  fun testAnswer_elementXWithPositionY_yValueDoesNotMatch() {
-    val inputs = mapOf("y" to NON_NEGATIVE_VALUE_0, "x" to VALID_CONTENT_ID_2)
+  fun testAnswer_elementXBeforeElementY_orderedCorrectly() {
+    val inputs = mapOf("x" to VALID_CONTENT_ID_1, "y" to VALID_CONTENT_ID_2)
 
     val matches =
-      hasElementXAtPositionYRuleClassifier.matches(
-        answer = LIST_OF_SETS_OF_CONTENT_IDS,
-        inputs = inputs,
-        classificationContext = ClassificationContext()
-      )
-
-    assertThat(matches).isFalse()
-  }
-
-  @Test
-  fun testAnswer_elementXWithPositionY_bothMatchesCorrectly() {
-    val inputs = mapOf("y" to NON_NEGATIVE_VALUE_1, "x" to VALID_CONTENT_ID_2)
-
-    val matches =
-      hasElementXAtPositionYRuleClassifier.matches(
+      hasElementXBeforeElementYRuleClassifier.matches(
         answer = LIST_OF_SETS_OF_CONTENT_IDS,
         inputs = inputs,
         classificationContext = ClassificationContext()
@@ -213,7 +197,7 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
   }
 
   private fun setUpTestApplicationComponent() {
-    DaggerDragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest_TestApplicationComponent
+    DaggerDragDropSortInputHasElementXBeforeElementYClassifierProviderTest_TestApplicationComponent
       .builder()
       .setApplication(ApplicationProvider.getApplicationContext()).build().inject(this)
   }
@@ -230,6 +214,6 @@ class DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest {
       fun build(): TestApplicationComponent
     }
 
-    fun inject(test: DragDropSortInputHasElementXAtPositionYRuleClassifierProviderTest)
+    fun inject(test: DragDropSortInputHasElementXBeforeElementYClassifierProviderTest)
   }
 }
